so viele errors

+++ web_application/src/controller/LecturerDeleteController.java
@@ -9,6 +9,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import helper.*;
 import models.*;
@@ -17,12 +18,21 @@ public class LecturerDeleteController extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
 
-		DBAccessJDBCSQLite db = new DBAccessJDBCSQLite();
-		ConvertObjectAndClass cac = new ConvertObjectAndClass();
-		db.connectTODB();
-		
-		db.delete(Integer.parseInt(req.getParameter("leid")), DBAccessJDBCSQLite.TableName.lecturer);
-		
-		getServletContext().getRequestDispatcher("/DeleteDozenten.jsp").forward(req, resp);
+		HttpSession session = req.getSession(true);
+		User user = (User) session.getAttribute("currentSessionUser");
+		if (user != null)
+		{
+			DBAccessJDBCSQLite db = new DBAccessJDBCSQLite();
+			ConvertObjectAndClass cac = new ConvertObjectAndClass();
+			db.connectTODB();
+			
+			db.delete(Integer.parseInt(req.getParameter("leid")), DBAccessJDBCSQLite.TableName.lecturer);
+			
+			getServletContext().getRequestDispatcher("/DeleteDozenten.jsp").forward(req, resp);
+		}
+		else 
+		{
+			getServletContext().getRequestDispatcher("/Login.jsp").forward(req, resp);
+		}
 	}
 }
