+++ web_application/src/controller/StudentDeleteController.java
@@ -9,6 +9,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import helper.*;
 import models.*;
@@ -17,19 +18,28 @@ public class StudentDeleteController extends HttpServlet{
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
 
-		DBAccessJDBCSQLite db = new DBAccessJDBCSQLite();
-		ConvertObjectAndClass cac = new ConvertObjectAndClass();
-		db.connectTODB();
-
-		if(db.delete(Integer.parseInt(req.getParameter("stid")), DBAccessJDBCSQLite.TableName.student))
+		HttpSession session = req.getSession(true);
+		User user = (User) session.getAttribute("currentSessionUser");
+		if (user != null)
 		{
-			req.setAttribute("info", "");
+			DBAccessJDBCSQLite db = new DBAccessJDBCSQLite();
+			ConvertObjectAndClass cac = new ConvertObjectAndClass();
+			db.connectTODB();
+	
+			if(db.delete(Integer.parseInt(req.getParameter("stid")), DBAccessJDBCSQLite.TableName.student))
+			{
+				req.setAttribute("info", "");
+			}
+			else
+			{
+				req.setAttribute("info", "");
+			}
+			
+			getServletContext().getRequestDispatcher("/DeleteStudenten.jsp").forward(req, resp);
 		}
-		else
+		else 
 		{
-			req.setAttribute("info", "");
+			getServletContext().getRequestDispatcher("/Login.jsp").forward(req, resp);
 		}
-		
-		getServletContext().getRequestDispatcher("/DeleteStudenten.jsp").forward(req, resp);
 	}
 }
